+++ server/core/src/main/java/dev/slimevr/reset/ResetHandler.kt
@@ -6,11 +6,11 @@ import java.util.function.Consumer
 class ResetHandler {
 	private val listeners: MutableList<ResetListener> = CopyOnWriteArrayList()
 
-	fun sendStarted(resetType: Int, bodyParts: List<Int>? = null, progress: Int = 0, duration: Int = 0) {
+	fun sendStarted(resetType: UByte, bodyParts: List<UByte>? = null, progress: Int = 0, duration: Int = 0) {
 		this.listeners.forEach { listener: ResetListener -> listener.onStarted(resetType, bodyParts, progress, duration) }
 	}
 
-	fun sendFinished(resetType: Int, bodyParts: List<Int>? = null, duration: Int) {
+	fun sendFinished(resetType: UByte, bodyParts: List<UByte>? = null, duration: Int) {
 		this.listeners.forEach { listener: ResetListener -> listener.onFinished(resetType, bodyParts, duration) }
 	}
