+++ server/core/src/main/java/dev/slimevr/reset/ResetListener.kt
@@ -1,7 +1,7 @@
 package dev.slimevr.reset
 
 interface ResetListener {
-	fun onStarted(resetType: Int, bodyParts: List<Int>? = null, progress: Int, duration: Int)
+	fun onStarted(resetType: UByte, bodyParts: List<UByte>? = null, progress: Int, duration: Int)
 
-	fun onFinished(resetType: Int, bodyParts: List<Int>? = null, duration: Int)
+	fun onFinished(resetType: UByte, bodyParts: List<UByte>? = null, duration: Int)
 }
